requirements.txt
pillow==11.1.0
qrcode==8.0
segno==1.6.1
